printf type-matching continues. again.

+++ dump.c
@@ -993,7 +993,7 @@ Perl_do_op_dump(pTHX_ I32 level, PerlIO *file, const OP *o)
                                       (label_flags & SVf_UTF8)));
    }
         Perl_dump_indent(aTHX_ level, file, "SEQ = %u\n",
-                                             cCOPo->cop_seq);
+                         (unsigned int)cCOPo->cop_seq);
 	break;
     case OP_ENTERLOOP:
 	Perl_dump_indent(aTHX_ level, file, "REDO ===> ");
